NM-106,210
Task 106 involved adjusting the filters in the homepage and overview so that when a task is marked complete in the overview, it disappears from the the overview and home page and only shows up in the history page marked as "complete" in the top right corner of the tasks. (note: you have to refresh the page or switch to another page for this effect to take place)

Task 210 involved just adjusting the filters in the home and overview page to ignore deleted tasks, and for the history page to show deleted tasks. For now, the only way to test this is to set the default deleted status of the schema task to true and create a task. This deleted tasks should then only show up in the history page and no where else.

The taskModel.js edit was just a small edit to remove unnecessary code.