JIRA/NM - 235, 236, 239, 249, 250
Added a new "Users" page where every user in the database is listed with their full name and email. Css is complete and polished for this page.

Deleted the Login/Signup buttons from the Navbar, and added the route to "Users" under the Menu drop down.

Added the css for ForgotPassword and ResetPassword pages, along with css for the actual reset password function on the Profile page. All css is consistent and neat.