NM - #264, 265, 266, 275
NM 264 - Designed a password requirement description box for the 'Change Password' and 'Sign Up' options to be used when a user changes their password or signs up as a new employee.

NM 265 - Implemented a notification if a user enters their current password incorrectly to change their password. Also, implemented the Email OTP success/fail indication design for the change password option. Now users can see all these options under the change password option.

NM 266 - Checked to ensure that all the incorrect credentials pop-ups and verified pop-ups are correctly positioned on each page. Also, moved the verification message to the same bar of admin details in the users' page under the admin login when the user clicks the up arrow to switch between 'Admin' and 'Employee' roles.

NM 275 - Implemented the color code for the 'In Progress' and 'Past Due' status boxes on the overview page. Additionally, separated the task number from the task title on the overview page. Now employees can easily see the 'In Progress' and 'Past Due' statuses in different colors.

A new library has been added to the frontend: validator. Please make sure to run 'npm install' after pulling my code to make sure that the frontend works properly. The npm library validator is used to validate the strength of the password on the frontend.