JIRA/NM 287,288,298
User page properly reflected on the Homepage as a user is deleted it shows up as User Deleted. You can then go through the overview page and go to edit task button and then remove the deleted user from the task. Each task in the home page now moves to In Progress or Past Due in real time. Meaning if it passes by the minute where its due the task will show in the past due location with proper change. The edit button in the overview page needs less than 10 seconds to appropriately update the time.