NM - #305, #306
NM #305 - Fixed the issue with the 'Forgot Password' link so that now users can request a password reset link as many times as they need, instead of it being sent only once.

NM #306 - Worked on the functionality of the 'Forgot Password' button, so now users can reset their password properly.